Events and news owner and admin permission. Postman collection updated

+++ backend/src/main/java/com/grupo6daw/lcdd_daw/controller/EventsRestController.java
@@ -2,6 +2,7 @@ package com.grupo6daw.lcdd_daw.controller;
 
 import java.io.IOException;
 import java.net.URI;
+import java.security.Principal;
 import java.sql.SQLException;
 import java.time.LocalDateTime;
 
@@ -9,6 +10,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -33,6 +35,7 @@ import com.grupo6daw.lcdd_daw.model.User;
 import com.grupo6daw.lcdd_daw.repository.UserRepository;
 import com.grupo6daw.lcdd_daw.service.EventService;
 import com.grupo6daw.lcdd_daw.service.ImageService;
+import com.grupo6daw.lcdd_daw.service.UserService;
 
 @RestController
 @RequestMapping("/api/v1/events")
@@ -43,6 +46,9 @@ public class EventsRestController {
 
     @Autowired
     private ImageService imageService;
+
+    @Autowired
+    private UserService userService;
 
     @Autowired
     private EventMapper eventMapper;
@@ -118,13 +124,18 @@ public class EventsRestController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<?> deleteEvent(@PathVariable long id, Authentication authentication) {
+    public ResponseEntity<?> deleteEvent(@PathVariable long id, Principal principal) throws SQLException {
 
-        long userId = Long.parseLong(authentication.getName());
-        User logged = userRepository.findById(userId).orElseThrow();
+        Long userId = Long.parseLong(principal.getName());
 
-        Event deleted = eventService.delete(id);
-        return ResponseEntity.ok(eventService.toDTO(deleted, logged));
+        Event deleted;
+
+        if (eventService.checkPermissions(eventService.findById(id), userService.findById(userId), false)) {
+            deleted = eventService.delete(id);
+        } else {
+            throw new AccessDeniedException("No tienes permiso para borrar este evento");
+        }
+        return ResponseEntity.ok(eventService.toDTO(deleted));
     }
 
     @DeleteMapping("/{eventId}/images/{imageId}")
@@ -141,18 +152,21 @@ public class EventsRestController {
     @PutMapping("/{id}")
     public ResponseEntity<?> replaceEvent(@PathVariable long id,
             @RequestBody EventDTO updatedEventDTO,
-            Authentication authentication) throws SQLException {
+            Principal principal) throws SQLException {
 
-        long userId = Long.parseLong(authentication.getName());
-        User logged = userRepository.findById(userId).orElseThrow();
-
+        Long userId = Long.parseLong(principal.getName());
         Event updatedEvent = eventMapper.toDomainFromFullDTO(updatedEventDTO);
 
         updatedEvent.setEventId(id);
         updatedEvent.setEventImage(eventService.findById(id).getEventImage());
-        eventService.save(updatedEvent);
 
-        return ResponseEntity.ok(eventService.toDTO(updatedEvent, logged));
+        if (eventService.checkPermissions(updatedEvent, userService.findById(userId), false)) {
+            eventService.save(updatedEvent);
+        } else {
+            throw new AccessDeniedException("No tienes permiso para editar este evento");
+        }
+
+        return ResponseEntity.ok(eventService.toDTO(updatedEvent));
     }
 
     @PutMapping("/{id}/images")
